frontend dev and update readme.md

+++ backend/controllers/product.controller.js
@@ -4,10 +4,10 @@ import Product from "../models/product.model.js";
 export const getProduct = async (req, res) => {
   try {
     const products = await Product.find({});
-    res.status(200).json(products);
+    res.status(200).json({ success: true, data: products });
   } catch (error) {
     console.error("Error: ", error.message);
-    res.status(500).json({ message: "Server Error" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
 
@@ -15,17 +15,19 @@ export const createProduct = async (req, res) => {
   const product = req.body;
 
   if (!product.name || !product.price || !product.image) {
-    return res.status(400).json({ message: "Please enter all fields" });
+    return res
+      .status(400)
+      .json({ success: false, message: "Please enter all fields" });
   }
 
   const newProduct = new Product(product);
 
   try {
     const createdProduct = await newProduct.save();
-    res.status(201).json(createdProduct);
+    res.status(201).json({ success: true, data: createdProduct });
   } catch (error) {
     console.error("Error: ", error.message);
-    res.status(500).json({ message: "Server Error" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
 
@@ -41,19 +43,25 @@ export const updateProduct = async (req, res) => {
     const updatedProduct = await Product.findByIdAndUpdate(id, product, {
       new: true,
     });
-    res.status(200).json(updatedProduct);
+    res.status(200).json({ success: true, data: updatedProduct });
   } catch (error) {
-    res.status(500).json({ message: "Server Error" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).send("Inavlid id");
+  }
+
   try {
     await Product.findByIdAndDelete(id);
-    res.status(200).json({ message: "Product deleted successfully" });
+    res
+      .status(200)
+      .json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
-    res.status(404).json({ message: "Product not found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
